Personal PC
A deep dive into understanding, cleaning, renaming, defining the data along with an attempt to derive valuable insights and uncover interesting patterns to achieve business excellence is what defines an Exploratory Data Analysis.
An exploratory data analysis enables you to analyze, investigate and summarize the main characteristics so we can try to find how we can use the data in order to solve complex problems, to spot anomalies, to test hypothesis and identify patterns to discover new business opportunities in future. In short, EDA is widely used by business/ data analyst to get a gist of the data structure and how we can use the data in a productive way.
Airbnb is an American San Francisco based company, providing an online marketplace for short- and long-term homestays and vacation since 2008.
The data here contains a list of features which will be used to determine whether the price has any effect on the other selective parameters, such as room type, duration of stay (minimum nights spent), reviews, room availability, etc. We are also going to search for the best prediction model to predict the price, going to conduct hypothesis testing to check if price affects the other features and also use the clustering techniques to get the optimal number of clusters in order to identify the opportunity statements respectively.